+++ ui/src/main/java/com/jin/ui/recipe/content/RecipeOverview.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.recipe.ui.content
+package com.jin.ui.recipe.content
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -28,7 +28,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
-import com.jin.honey.R
+import com.jin.ui.R
 import com.jin.ui.theme.PointColor
 
 @Composable
